Fiffi POS 17.06.2020
Scoreboard fast fertig + wieda wos bei layouts

+++ app/src/main/java/com/example/dartswithfriends/Match.java
@@ -4,11 +4,11 @@ import java.util.HashMap;
 import java.util.Map;
 
 public class Match {
-    HashMap<Player,Integer> game;
+    HashMap<Player,Double> game;
     Double longitude;
     Double latitude;
 
-    public Match(HashMap<Player, Integer> game, Double longitude, Double latitude) {
+    public Match(HashMap<Player, Double> game, Double longitude, Double latitude) {
         this.game = game;
         this.longitude = longitude;
         this.latitude = latitude;
