+++ app/src/main/java/com/example/dartswithfriends/ScoreBoardLvAdapter.java
@@ -52,7 +52,7 @@ public class ScoreBoardLvAdapter extends BaseAdapter {
             for( Player x : players){
                 //PLAYER 1
                 ((TextView) listItem.findViewById(R.id.scoreboard_Player1_txtView)).setText(x.name);
-                int av = match.game.get(x);
+                Double av = match.game.get(x);
                 String avs = String.valueOf(av);
                 ((TextView) listItem.findViewById(R.id.scoreboard_averageP1_txtView)).setText("Avg.: "+avs);
                 //PLAYER 2
@@ -78,14 +78,14 @@ public class ScoreBoardLvAdapter extends BaseAdapter {
                 //PLAYER 1
                 if(counter==0) {
                     ((TextView) listItem.findViewById(R.id.scoreboard_Player1_txtView)).setText(x.name);
-                    int av = match.game.get(x);
+                    Double av = match.game.get(x);
                     String avs = String.valueOf(av);
                     ((TextView) listItem.findViewById(R.id.scoreboard_averageP1_txtView)).setText("Avg.: "+avs);
                 }
                 //PLAYER 2
                 if(counter==1) {
                     ((TextView) listItem.findViewById(R.id.scoreboard_player2_txtView)).setText(x.name);
-                    int av = match.game.get(x);
+                    Double av = match.game.get(x);
                     String avs = String.valueOf(av);
                     ((TextView) listItem.findViewById(R.id.scoreboard_averageP2_txtView)).setText("Avg.: "+avs);
                 }
@@ -119,21 +119,21 @@ public class ScoreBoardLvAdapter extends BaseAdapter {
                 //PLAYER 1
                 if(counter==0) {
                     ((TextView) listItem.findViewById(R.id.scoreboard_Player1_txtView)).setText(x.name);
-                    int av = match.game.get(x);
+                    Double av = match.game.get(x);
                     String avs = String.valueOf(av);
                     ((TextView) listItem.findViewById(R.id.scoreboard_averageP1_txtView)).setText("Avg.: "+avs);
                 }
                 //PLAYER 2
                 if(counter==1) {
                     ((TextView) listItem.findViewById(R.id.scoreboard_player2_txtView)).setText(x.name);
-                    int av = match.game.get(x);
+                    Double av = match.game.get(x);
                     String avs = String.valueOf(av);
                     ((TextView) listItem.findViewById(R.id.scoreboard_averageP2_txtView)).setText("Avg.: "+avs);
                 }
                 //PLAYER 3
                 if(counter==2) {
                     ((TextView) listItem.findViewById(R.id.scoreboard_player3_txtView)).setText(x.name);
-                    int av = match.game.get(x);
+                    Double av = match.game.get(x);
                     String avs = String.valueOf(av);
                     ((TextView) listItem.findViewById(R.id.scoreboard_averageP3_txtView)).setText("Avg.: "+avs);
                 }
@@ -158,28 +158,28 @@ public class ScoreBoardLvAdapter extends BaseAdapter {
                 //PLAYER 1
                 if(counter==0) {
                     ((TextView) listItem.findViewById(R.id.scoreboard_Player1_txtView)).setText(x.name);
-                    int av = match.game.get(x);
+                    Double av = match.game.get(x);
                     String avs = String.valueOf(av);
                     ((TextView) listItem.findViewById(R.id.scoreboard_averageP1_txtView)).setText("Avg.: "+avs);
                 }
                 //PLAYER 2
                 if(counter==1) {
                     ((TextView) listItem.findViewById(R.id.scoreboard_player2_txtView)).setText(x.name);
-                    int av = match.game.get(x);
+                    Double av = match.game.get(x);
                     String avs = String.valueOf(av);
                     ((TextView) listItem.findViewById(R.id.scoreboard_averageP2_txtView)).setText("Avg.: "+avs);
                 }
                 //PLAYER 3
                 if(counter==2) {
                     ((TextView) listItem.findViewById(R.id.scoreboard_player3_txtView)).setText(x.name);
-                    int av = match.game.get(x);
+                    Double av = match.game.get(x);
                     String avs = String.valueOf(av);
                     ((TextView) listItem.findViewById(R.id.scoreboard_averageP3_txtView)).setText("Avg.: "+avs);
                 }
                 //PLAYER 4
                 if(counter==3) {
                     ((TextView) listItem.findViewById(R.id.scoreboard_player4_txtView)).setText(x.name);
-                    int av = match.game.get(x);
+                    Double av = match.game.get(x);
                     String avs = String.valueOf(av);
                     ((TextView) listItem.findViewById(R.id.scoreboard_averageP4_txtView)).setText("Avg.: "+avs);
                 }
